LayoutTests/media/modern-media-controls/media-controller/media-controller-resize.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;"></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>MediaController</code> resizing behavior.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

let numberOfFrames = 0;
scheduler.frameDidFire = function()
{
    numberOfFrames++;

    if (numberOfFrames == 1) {
        debug("Checking initial size");
        shouldBe("mediaController.controls.width", "320");
        shouldBe("mediaController.controls.height", "240");
        
        debug("");
        debug("Resizing to 400x300");
        media.style.width = "400px";
        media.style.height = "300px";
    } else {
        shouldBe("mediaController.controls.width", "400");
        shouldBe("mediaController.controls.height", "300");

        debug("");
        shadowRoot.host.remove();
        media.remove();
        finishJSTest();
    }
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
